check authorization update user, book

+++ src/main/java/com/kimho/book/service/impl/BookServiceImpl.java
@@ -3,8 +3,11 @@ package com.kimho.book.service.impl;
 import com.kimho.book.converter.BookToDto;
 import com.kimho.book.converter.DtoToBook;
 import com.kimho.book.exception.NotFoundException;
+import com.kimho.book.exception.UnauthorizedException;
 import com.kimho.book.model.dao.Book;
+import com.kimho.book.model.dao.User;
 import com.kimho.book.model.dto.BookDto;
+import com.kimho.book.model.dto.BookUpdate;
 import com.kimho.book.repository.BookRepository;
 import com.kimho.book.service.BooksService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -12,14 +15,19 @@ import org.springframework.stereotype.Service;
 
 import java.util.Date;
 import java.util.List;
+
 @Service
-public class BookServiceImpl implements BooksService<BookDto> {
+public class BookServiceImpl implements BooksService<BookDto, BookUpdate> {
     @Autowired
     private BookRepository bookRepository;
     @Autowired
     private BookToDto bookToDto;
     @Autowired
     private DtoToBook dtoToBook;
+    @Autowired
+    private UserServiceImpl userService;
+    private final User myUser = userService.getMyUser();
+
     @Override
     public List<BookDto> getAll() {
         List<Book> books = bookRepository.findAll();
@@ -32,21 +40,32 @@ public class BookServiceImpl implements BooksService<BookDto> {
         return bookToDto.convert(bookRepository.findById(id).get());
     }
 
+    private Book findById(long id) {
+        verifyBookIsExist(id);
+        return bookRepository.findById(id).get();
+    }
+
     @Override
     public BookDto post(BookDto bookDto) {
         Book book = dtoToBook.convert(bookDto);
+        if (myUser.isAdmin() || myUser.isSuperAdmin()) {
+            book.setEnabled(true);
+        }
         BookDto bookResponse = bookToDto.convert(bookRepository.save(book));
         return bookResponse;
     }
 
     @Override
-    public BookDto put(BookDto bookDto, long id) {
+    public BookDto put(BookUpdate bookEdition, long id) {
         verifyBookIsExist(id);
         Book book = bookRepository.findById(id).get();
-        book.setTitle(bookDto.getTitle());
-        book.setAuthor(bookDto.getAuthor());
-        book.setDescription(bookDto.getDescription());
-        book.setImage(bookDto.getImage());
+        if (!checkAuthorizationEditBook(book)) {
+            throw new UnauthorizedException("Unauthorized");
+        }
+        book.setTitle(bookEdition.getTitle());
+        book.setAuthor(bookEdition.getAuthor());
+        book.setDescription(bookEdition.getDescription());
+        book.setImage(bookEdition.getImage());
         book.setUpdatedAt(new Date());
         return bookToDto.convert(bookRepository.save(book));
     }
@@ -54,17 +73,47 @@ public class BookServiceImpl implements BooksService<BookDto> {
     @Override
     public void delete(long id) {
         verifyBookIsExist(id);
+        if (!checkAuthorizationEditBook(findById(id))) {
+            throw new UnauthorizedException("Unauthorized");
+        }
         bookRepository.deleteById(id);
     }
 
-    public BookDto enabled(long id){
+    public BookDto enable(long id) {
         Book book = bookRepository.getById(id);
+        if (!checkAuthorizationEnableBook(book)){
+            throw new UnauthorizedException("Unauthorized");
+        }
         book.setEnabled(!book.isEnabled());
         return bookToDto.convert(bookRepository.save(book));
     }
 
-    public void verifyBookIsExist(long id){
-        if (!bookRepository.existsById(id)){
+    private boolean checkAuthorizationEditBook(Book book) {
+        User userCreateBook = book.getUser();
+        if (userCreateBook.getId() == myUser.getId()) {
+            return true;
+        } else if (myUser.isSuperAdmin()) {
+            return true;
+        } else if (userCreateBook.isUser() && myUser.isAdmin()) {
+            return true;
+        }
+        return false;
+    }
+
+    private boolean checkAuthorizationEnableBook(Book book) {
+        User userCreateBook = book.getUser();
+        if (myUser.isSuperAdmin()) {
+            return true;
+        } else if (myUser.isAdmin() && userCreateBook.isUser()) {
+            return true;
+        } else if (myUser.isAdmin() && (myUser.getId() == userCreateBook.getId())) {
+            return true;
+        }
+        return false;
+    }
+
+    public void verifyBookIsExist(long id) {
+        if (!bookRepository.existsById(id)) {
             throw new NotFoundException("Book not found");
         }
     }
